+++ ast/tidb.go
@@ -17,13 +17,15 @@
 package ast
 
 import (
+	"fmt"
+
 	"github.com/XiaoMi/soar/common"
 
+	json "github.com/CorgiMan/json2"
 	"github.com/kr/pretty"
 	"github.com/pingcap/parser"
 	"github.com/pingcap/parser/ast"
-
-	json "github.com/CorgiMan/json2"
+	"github.com/tidwall/gjson"
 
 	// for pingcap parser
 	_ "github.com/pingcap/tidb/types/parser_driver"
@@ -66,4 +68,82 @@ func StmtNode2JSON(sql, charset, collation string) string {
 		}
 	}
 	return str
+}
+
+// SchemaMetaInfo get used database, table name from SQL
+func SchemaMetaInfo(sql string, defaultDatabase string) []string {
+	var tables []string
+	tree, err := TiParse(sql, "", "")
+	if err != nil {
+		return tables
+	}
+
+	jsonString := StmtNode2JSON(sql, "", "")
+
+	for _, node := range tree {
+		switch n := node.(type) {
+		case *ast.UseStmt:
+			tables = append(tables, fmt.Sprintf("`%s`.`dual`", n.DBName))
+		case *ast.InsertStmt, *ast.SelectStmt, *ast.UnionStmt, *ast.UpdateStmt, *ast.DeleteStmt:
+			// DML/DQL: INSERT, SELECT, UPDATE, DELETE
+			for _, tableRef := range common.JSONFind(jsonString, "TableRefs") {
+				for _, source := range common.JSONFind(tableRef, "Source") {
+					database := gjson.Get(source, "Schema.O")
+					table := gjson.Get(source, "Name.O")
+					if database.String() == "" {
+						if table.String() != "" {
+							tables = append(tables, fmt.Sprintf("`%s`.`%s`", defaultDatabase, table.String()))
+						}
+					} else {
+						if table.String() != "" {
+							tables = append(tables, fmt.Sprintf("`%s`.`%s`", database.String(), table.String()))
+						} else {
+							tables = append(tables, fmt.Sprintf("`%s`.`dual`", database.String()))
+						}
+					}
+				}
+			}
+		case *ast.DropTableStmt:
+			// DDL: DROP TABLE|VIEW
+			schemas := common.JSONFind(jsonString, "Tables")
+			for _, tabs := range schemas {
+				for _, table := range gjson.Parse(tabs).Array() {
+					db := gjson.Get(table.String(), "Schema.O")
+					tb := gjson.Get(table.String(), "Name.O")
+					if db.String() == "" {
+						if tb.String() != "" {
+							tables = append(tables, fmt.Sprintf("`%s`.%s`", defaultDatabase, tb.String()))
+						}
+					} else {
+						if tb.String() != "" {
+							tables = append(tables, fmt.Sprintf("`%s`.`%s`", db.String(), tb.String()))
+						}
+					}
+				}
+			}
+		case *ast.DropDatabaseStmt, *ast.CreateDatabaseStmt:
+			// DDL: DROP|CREATE DATABASE
+			schemas := common.JSONFind(jsonString, "Name")
+			for _, schema := range schemas {
+				tables = append(tables, fmt.Sprintf("`%s`.`dual`", schema))
+			}
+		default:
+			// DDL: CREATE TABLE|DATABASE|INDEX|VIEW, DROP INDEX
+			schemas := common.JSONFind(jsonString, "Table")
+			for _, table := range schemas {
+				db := gjson.Get(table, "Schema.O")
+				tb := gjson.Get(table, "Name.O")
+				if db.String() == "" {
+					if tb.String() != "" {
+						tables = append(tables, fmt.Sprintf("`%s`.%s`", defaultDatabase, tb.String()))
+					}
+				} else {
+					if tb.String() != "" {
+						tables = append(tables, fmt.Sprintf("`%s`.`%s`", db.String(), tb.String()))
+					}
+				}
+			}
+		}
+	}
+	return common.RemoveDuplicatesItem(tables)
 }