+++ ast/tidb_test.go
@@ -56,3 +56,33 @@ func TestStmtNode2JSON(t *testing.T) {
 	}
 	common.Log.Debug("Exiting function: %s", common.GetFunctionName())
 }
+
+func TestSchemaMetaInfo(t *testing.T) {
+	common.Log.Debug("Entering function: %s", common.GetFunctionName())
+	sqls := []string{
+		"use world_x;",
+		"select 1;",
+		"syntax error case",
+		"select * from ta join tb using (id)",
+		"select * from ta, tb limit 1",
+		"drop table tb",
+		"drop table db.tb",
+		"drop database db",
+		"create database db",
+		"create index idx_col on tbl (col)",
+		"DROP INDEX idx_col on tbl",
+	}
+	// fmt.Println(sqls[len(sqls)-1])
+	// fmt.Println(SchemaMetaInfo(sqls[len(sqls)-1], "sakila"))
+	// return
+	err := common.GoldenDiff(func() {
+		for _, sql := range append(sqls, common.TestSQLs...) {
+			fmt.Println(sql)
+			fmt.Println(SchemaMetaInfo(sql, "sakila"))
+		}
+	}, t.Name(), update)
+	if nil != err {
+		t.Fatal(err)
+	}
+	common.Log.Debug("Exiting function: %s", common.GetFunctionName())
+}